main part over

+++ model/question.go
@@ -3,11 +3,11 @@ package model
 import "encoding/json"
 
 type Question struct {
-	Id   int    `json:"id" db:"id"`
+	Id   int64  `json:"id" db:"id"`
 	Word string `json:"word" db:"word"`
 }
 
-func NewQuestion(id int, word string) *Question {
+func NewQuestion(id int64, word string) *Question {
 	return &Question{id, word}
 }
